requirements.txt
langextract>=1.1.1
python-dotenv>=1.0.1
streamlit>=1.32.0
pypdf>=5.1.0
python-docx>=1.1.2
pdfplumber>=0.11.4
pytesseract>=0.3.10
pdf2image>=1.17.0